+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "recyclerlistview-gridlayoutmanager",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "Grid Layout Manager built on top of RecyclerListView!",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -34,7 +34,7 @@
     "url": "https://github.com/muskeinsingh/recyclerlistview-gridlayoutmanager/issues"
   },
   "homepage": "https://github.com/muskeinsingh/recyclerlistview-gridlayoutmanager",
-  "peerDependencies": {
-    "recyclerlistview": ">= 1.4.0"
+  "dependencies": {
+    "recyclerlistview": "1.4.0-beta.6"
   }
 }
