+++ scripts/build.sh
@@ -6,7 +6,10 @@ tslint 'src/**/*.{ts,tsx}'
 
 echo "Build started..."
 echo "Removing old builds..."
+rm -rf node_modules
 rm -rf dist
+echo "Installing dependencies"
+npm install
 
 echo "TSC: Building package..."
 tsc --outDir dist
