+++ src/index.ts
@@ -1,10 +1,7 @@
-import { GridLayoutProvider, Dimension } from "./LayoutProvider";
-import { GridLayoutManager, Point, Layout } from "./LayoutManager";
+import GridLayoutProvider from "./GridLayoutProvider";
+import GridLayoutManager from "./GridLayoutManager";
 
 export {
     GridLayoutProvider,
     GridLayoutManager,
-    Dimension,
-    Point,
-    Layout,
 };
